Move types to types folder
