+++ src/components/stories/StoryPreview.tsx
@@ -3,7 +3,7 @@ import Card from '@mui/material/Card';
 import CardContent from '@mui/material/CardContent';
 import CardMedia from '@mui/material/CardMedia';
 import Typography from '@mui/material/Typography';
-import { Story } from '../../utils/types';
+import { Story } from '../../types';
 import { Box, styled, useTheme } from '@mui/material';
 import parse from 'html-react-parser';
 //import { convertDateToDotFormat } from '../../utils/date';
